[BSK 676] Committing star tracker fixes and some minor UKF updates that are likely going to get tossed overboard.
Former-commit-id: bf54e52f2ebbfb9a33e16d904822ac480018d219
Former-commit-id: 3f1a3dbd32fbce6c34c296b776410ab211c70a54

+++ FswAlgorithms/fswMessages/stAttFswMsg.h
@@ -23,7 +23,7 @@
 
 /*! @brief Output structure for ST attitude measurement in vehicle body frame*/
 typedef struct {
-    double timeTag;              /*!< [s] Vehicle time code associated with measurement*/
+    uint64_t timeTag;              /*!< [ns] Vehicle time code associated with measurement*/
     double MRP_BdyInrtl[3];      /*!< [-] MRP estimate of inertial to body transformation*/
 }STAttFswMsg;
